+++ MoneyApp/MoneyApp/Server/Repositories/IncomeCategoryRepository.cs
@@ -6,38 +6,58 @@ namespace MoneyApp.Server.Repositories;
 public class IncomeCategoryRepository : IIncomeCategoryRepository
 {
     private readonly DataContext _context;
+    private readonly IUserRepository _userRepository;
 
-    public IncomeCategoryRepository(DataContext context)
+
+    public IncomeCategoryRepository(DataContext context, IUserRepository userRepository)
     {
         _context = context;
+        _userRepository = userRepository;
     }
 
     public async Task<IEnumerable<IncomeCategory>> GetAllIncomeCategories()
     {
-        return await _context.IncomeCategories.ToListAsync();
+        var currentUser = _userRepository.GetCurrentUser();
+        return await _context.IncomeCategories
+            .Where(x => x.UserId == currentUser)
+            .ToListAsync();
     }
 
     public async Task<IncomeCategory> GetIncomeCategoryById(int id)
     {
-        return await _context.IncomeCategories.FirstOrDefaultAsync(x => x.Id == id);
+        var currentUser = _userRepository.GetCurrentUser();
+        return await _context.IncomeCategories
+            .Where(x => x.UserId == currentUser)
+            .FirstOrDefaultAsync(x => x.Id == id);
     }
 
     public async Task CreateIncomeCategory(IncomeCategory incomeCategory)
     {
+        incomeCategory.UserId = _userRepository.GetCurrentUser();
         _context.IncomeCategories.Add(incomeCategory);
         await _context.SaveChangesAsync();
     }
 
     public async Task UpdateIncomeCategory(IncomeCategory incomeCategory)
     {
-        _context.IncomeCategories.Update(incomeCategory);
-        await _context.SaveChangesAsync();
+        var currentUser = _userRepository.GetCurrentUser();
+
+        if (incomeCategory.UserId == currentUser)
+        {
+            _context.IncomeCategories.Update(incomeCategory);
+            await _context.SaveChangesAsync();
+        }
     }
 
     public async Task DeleteIncomeCategory(int id)
     {
+        var currentUser = _userRepository.GetCurrentUser();
         var incomeCategory = await GetIncomeCategoryById(id);
-        _context.IncomeCategories.Remove(incomeCategory);
-        await _context.SaveChangesAsync();
+
+        if (incomeCategory.UserId == currentUser)
+        {
+            _context.IncomeCategories.Remove(incomeCategory);
+            await _context.SaveChangesAsync();
+        }
     }
 }
